requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==3.2.23
djangorestframework==3.14.0
importlib-metadata==4.13.0
kombu==5.2.4
prompt-toolkit==3.0.43
python-dotenv==0.21.1
pytz==2023.4
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.15.0
